Adapting control tasks to the change in ControlLoops message.

+++ src/Control/AUV/Allocator/Task.cpp
@@ -68,15 +68,15 @@ namespace Control
         unsigned m_spos_ent;
         //! Servo positions
         float m_servo_pos[c_fins];
-        //! Control loops last reference time
-        float m_scope_ref;
+        //! Control loops last reference
+        uint32_t m_scope_ref;
         //! Task arguments.
         Arguments m_args;
 
         Task(const std::string& name, Tasks::Context& ctx):
           Tasks::Task(name, ctx),
           m_braking(false),
-          m_scope_ref(0.0)
+          m_scope_ref(0)
         {
           param(DTR_RT("Maximum Fin Rotation"), m_args.max_fin_rot)
           .defaultValue("25.0")
