+++ src/Control/AUV/Diving/Task.cpp
@@ -104,8 +104,8 @@ namespace Control
         bool m_last_try;
         //! Braking flag
         bool m_braking;
-        //! Control loops last reference time
-        float m_scope_ref;
+        //! Control loops last reference
+        uint32_t m_scope_ref;
         //! Task arguments.
         Arguments m_args;
 
@@ -115,7 +115,7 @@ namespace Control
           m_counter_solo(NULL),
           m_counter_step(NULL),
           m_braking(false),
-          m_scope_ref(0.0)
+          m_scope_ref(0)
         {
           param("Depth Tolerance", m_args.depth_tol)
           .defaultValue("0.5")
